+++ iModelCore/ECDb/ECDb/ECDbImpl.cpp
@@ -266,6 +266,7 @@ void ECDb::Impl::RegisterECSqlPragmas() const
     GetPragmaManager().Register(PragmaDbList::Create());
     GetPragmaManager().Register(PragmaCheckECSqlWriteValues::Create());
     GetPragmaManager().Register(PragmaECSqlVersion::Create());
+    GetPragmaManager().Register(PragmaSqliteSql::Create());
     }
 
 //--------------------------------------------------------------------------------------