+++ iModelCore/ECDb/ECDb/ECSql/ECSqlPragmas.h
@@ -69,6 +69,17 @@ struct PragmaECSqlVersion : PragmaManager::GlobalHandler {
 };
 
 //=======================================================================================
+// @bsiclass
+//+===============+===============+===============+===============+===============+======
+struct PragmaSqliteSql : PragmaManager::GlobalHandler {
+    PragmaSqliteSql() : GlobalHandler("sqlite_sql", "return underlying SQLite SQL") {}
+    ~PragmaSqliteSql() {}
+    virtual DbResult Read(PragmaManager::RowSet&, ECDbCR, const PragmaVal&, const PragmaManager::OptionsMap&) override;
+    virtual DbResult Write(PragmaManager::RowSet&, ECDbCR, const PragmaVal&, const PragmaManager::OptionsMap&) override;
+    static std::unique_ptr<PragmaManager::Handler> Create() { return std::make_unique<PragmaSqliteSql>(); }
+};
+
+//=======================================================================================
 // @bsiclass PragmaChecksum
 //+===============+===============+===============+===============+===============+======
 struct PragmaChecksum : PragmaManager::GlobalHandler {